WebScrapping of BookStore Data
The goal of this project is  to scrape multiple web pages using BeautifulSoup and export the results to a CSV file using Pandas Library. We use books.toscrape.com as the web scraping playground and get access to the alt attribute of the images and the name of the classes.